backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface-hub==0.20.0
openai==1.3.5
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2

